+++ standard/CHANGELOG.md
@@ -1,6 +1,11 @@
 
 ## v7.0.0-beta.8
 
+- fix: vanilla CSS themes, sizing relative to component rather than html tag, and uses em units
+
+
+## v7.0.0-beta.8
+
 - feature: `colorScheme` prop on component
 - feature: `resourceCell*` hooks now receive `.field` in data
 - feature: new Classic theme CSS variables: